% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractmatcru.R
\name{extractareacru}
\alias{extractareacru}
\title{Function for extracting the climate data of an area from CRU datasets}
\usage{
extractareacru(file, lonmin, lonmax, latmin, latmax, var, year, month, precision)
}
\arguments{
\item{file}{The CRU dataset file, usually with ".dat.nc" extension}

\item{lonmin}{The minimum longitude value of the rectangular area}

\item{lonmax}{The maximum longitude value of the rectangular area}

\item{latmin}{The minimum latitude value of the rectangular area}

\item{latmax}{The maximum latitude value of the rectangular area}

\item{var}{The variable from the CRU dataset that we would like to obtain,
could be "tmp" for temperature, "pre" for precipitation, and so on.}

\item{year}{A single given year, must be within the span measured in the CRU dataset}

\item{month}{A single given month, must be from the Gregorian calendar and can be written either in English ("January", "February", etc.) or number (1, 2, 3, etc.).}
}
\value{
A matrix with the desired climate values (precipitation, temperature, etc.) of
the rectangular region on the requested year and month. Rows refer to latitude and columns
refer to longitude.
}
\description{
A wrapper for extractcru() function,
useful for extracting climate date of
a rectangular area from CRU datasets.
}
\examples{
# Selected time
year <- 2010
month <- "December"

# Selected area: a rectangular region of the Central Java island, Indonesia
lonmin <- 109.0
lonmax <- 109.2
latmin <- -7.4
latmax <- -6.9

# Climate parameter
var <- "pre"

# CRU dataset
file <- "cru_ts4.02.1901.2017.tmp.dat.nc"

# Coordinate precision of 1 digit behind decimal
precision <- 1

# Matrix with results
matReg <- extractareacru(file, lonmin, lonmax, latmin, latmax,
          var, year, month, precision)
}
